# contracts/v1-openapi.yaml
openapi: 3.1.0
info:
  title: LogiFlow API
  version: 1.0.0
  description: |
    LogiFlow logistics management API. Single source of truth for backend and frontend.
    Base path: /api/v1
  contact:
    name: LogiFlow
    url: https://logiflow.dev
  license:
    name: Proprietary
servers:
  - url: https://localhost:5001/api/v1
    description: Local development
  - url: https://logiflow.example.com/api/v1
    description: Production (single-node v1)
tags:
  - { name: System, description: Health and metadata }
  - { name: Auth, description: Authentication and token management }
  - { name: Warehouses, description: Warehouse master data }
  - { name: Vehicles, description: Vehicle master data }
  - { name: Drivers, description: Driver master data }
  - { name: Shipments, description: Shipment lifecycle }
  - { name: Routes, description: Route planning and assignment }
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Issue via POST /auth/login. Required on every endpoint except
        /auth/login and /auth/refresh.
  schemas:
    HealthResponse:
      type: object
      properties:
        status: { type: string, example: ok }
        service: { type: string, example: LogiFlow }
        timestamp: { type: string, format: date-time, example: 2026-09-18T00:00:00Z }
      required: [status, service, timestamp]
    ProblemDetails:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri, example: https://logiflow.dev/errors/validation }
        title: { type: string, example: Validation failed }
        status: { type: integer, format: int32, example: 400 }
        detail: { type: string, nullable: true }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
        traceId: { type: string, nullable: true, example: "00-abc123" }
    WarehouseRequest:
      type: object
      required: [name, address]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          example: Central DC
        address:
          type: string
          minLength: 1
          maxLength: 500
          example: 12 Industrial Rd, Rotterdam
        latitude: { type: number, format: double, minimum: -90, maximum: 90, example: 51.9244 }
        longitude: { type: number, format: double, minimum: -180, maximum: 180, example: 4.4777 }
    WarehouseResponse:
      type: object
      required: [id, name, address, createdAt]
      properties:
        id: { type: integer, format: int64, example: 1 }
        name: { type: string, example: Central DC }
        address: { type: string, example: 12 Industrial Rd, Rotterdam }
        latitude: { type: number, format: double, nullable: true, example: 51.9244 }
        longitude: { type: number, format: double, nullable: true, example: 4.4777 }
        createdAt: { type: string, format: date-time, example: 2026-09-18T08:00:00Z }
    PagedResponse:
      type: object
      description: Envelope for all list endpoints (max pageSize 100, default 25).
      required: [items, page, pageSize, totalCount, totalPages]
      properties:
        items:
          type: array
          items: {}
          description: Array of resource objects (schema defined per endpoint).
        page: { type: integer, minimum: 1, default: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 100, default: 25 }
        totalCount: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
  responses:
    ValidationProblem:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Forbidden:
      description: Authenticated but not authorized for this role/resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
  # Conventions per 05-api-contract-standards.md §Auth + §Status transitions
  x-role-sets:
    Admin: [Admin]
    Dispatcher: [Dispatcher]
    Driver: [Driver]
    Viewer: [Viewer]
  x-conventions:
    resource-naming: plural nouns
    status-transitions: POST /{resource}/{id}/status-transitions with { "toStatus":"...", "note":"..." }
paths:
  /health:
    get:
      tags: [System]
      summary: Service health
      security: []
      x-roles: [Admin, Dispatcher, Driver, Viewer]
      operationId: getHealth
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HealthResponse' }
  /warehouses:
    get:
      tags: [Warehouses]
      summary: List warehouses (paged)
      description: >
        Returns a paged list of warehouses. Query params: page (default 1),
        pageSize (default 25, max 100), q (name contains, case-insensitive).
      security: []
      x-roles: [Admin, Dispatcher, Viewer]
      operationId: listWarehouses
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1, minimum: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 25, minimum: 1, maximum: 100 } }
        - { name: q, in: query, description: "Filter by name (contains, case-insensitive)", schema: { type: string } }
      responses:
        "200":
          description: Paged warehouse list
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/PagedResponse' }
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/WarehouseResponse' }
    post:
      tags: [Warehouses]
      summary: Create warehouse
      description: Creates a warehouse. Name and address are required; coordinates optional but range-validated.
      security: []
      x-roles: [Admin, Dispatcher]
      operationId: createWarehouse
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WarehouseRequest' }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WarehouseResponse' }
        "400": { $ref: '#/components/responses/ValidationProblem' }
  /warehouses/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      tags: [Warehouses]
      summary: Get warehouse by id
      security: []
      x-roles: [Admin, Dispatcher, Viewer]
      operationId: getWarehouse
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WarehouseResponse' }
        "404": { $ref: '#/components/responses/ValidationProblem' }
    put:
      tags: [Warehouses]
      summary: Update warehouse (full update)
      security: []
      x-roles: [Admin, Dispatcher]
      operationId: updateWarehouse
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WarehouseRequest' }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WarehouseResponse' }
        "400": { $ref: '#/components/responses/ValidationProblem' }
        "404": { $ref: '#/components/responses/ValidationProblem' }
    delete:
      tags: [Warehouses]
      summary: Delete warehouse
      description: Hard delete in v1 (no soft-delete column). Later restricted once shipments reference warehouses.
      security: []
      x-roles: [Admin]
      operationId: deleteWarehouse
      responses:
        "204": { description: Deleted }
        "404": { $ref: '#/components/responses/ValidationProblem' }
  /auth/login:
    post:
      tags: [Auth]
      summary: Issue access + refresh tokens
      security: []
      x-roles: [Admin, Dispatcher, Driver, Viewer]
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email, example: dana@logiflow.dev }
                password: { type: string, format: password }
      responses:
        "200":
          description: Tokens issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  refresh_token: { type: string }
                  expires_in: { type: integer }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      security: []
      x-roles: [Admin, Dispatcher, Driver, Viewer]
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  expires_in: { type: integer }
security:
  - BearerAuth: []